requirements.txt
flask==2.2.2
transformers==4.27.4
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.1
gunicorn==20.1.0
PyYAML==6.0
pandas==1.5.3
scikit-learn==1.2.2
tqdm==4.65.0
joblib==1.2.0
mlflow==2.4.1
langdetect==1.0.9
fasttext==0.9.2
faiss-cpu==1.7.4
shap==0.41.0
python-jose==3.3.0
elasticsearch==8.7.0
psycopg2==2.9.6
uvloop==0.17.0
fastapi==0.95.2
uvicorn==0.22.0
graphene==3.1.1